+++ fabric/src/main/java/io/github/mortuusars/horseman/fabric/HorsemanFabric.java
@@ -19,7 +19,8 @@ public class HorsemanFabric implements ModInitializer {
     public void onInitialize() {
         Horseman.init();
 
-        NeoForgeConfigRegistry.INSTANCE.register(Horseman.ID, net.neoforged.fml.config.ModConfig.Type.COMMON, Config.Common.SPEC);
+        NeoForgeConfigRegistry.INSTANCE.register(Horseman.ID, ModConfig.Type.SERVER, Config.Server.SPEC);
+        NeoForgeConfigRegistry.INSTANCE.register(Horseman.ID, ModConfig.Type.COMMON, Config.Common.SPEC);
         NeoForgeConfigRegistry.INSTANCE.register(Horseman.ID, ModConfig.Type.CLIENT, Config.Client.SPEC);
 
         Horseman.Advancements.register();