+++ neoforge/src/main/java/io/github/mortuusars/horseman/neoforge/HorsemanNeoForge.java
@@ -16,6 +16,7 @@ public class HorsemanNeoForge {
     public HorsemanNeoForge(ModContainer container) {
         Horseman.init();
 
+        container.registerConfig(ModConfig.Type.SERVER, Config.Server.SPEC);
         container.registerConfig(ModConfig.Type.COMMON, Config.Common.SPEC);
         container.registerConfig(ModConfig.Type.CLIENT, Config.Client.SPEC);